+++ actix-http/src/h1/dispatcher.rs
@@ -842,10 +842,13 @@ where
                  A Request head too large to parse is only checked on
                  httparse::Status::Partial condition.
                 */
-                if this.payload.is_none() {
+                match this.payload {
                     /*
                     When dispatcher has a payload the responsibility of
                     wake up it would be shift to h1::payload::Payload.
+                    Unless the payload is ready to read, in which case
+                    it might not have access to the waker and could
+                    result in the dispatcher getting stuck until timeout.
 
                     Reason:
                     Self wake up when there is payload would waste poll
@@ -857,9 +860,9 @@ where
                     At this case read_buf could always remain beyond
                     MAX_BUFFER_SIZE and self wake up would be busy poll
                     dispatcher and waste resource.
-
                     */
-                    cx.waker().wake_by_ref();
+                    Some(ref p) if p.need_read(cx) != PayloadStatus::Read => {}
+                    _ => cx.waker().wake_by_ref(),
                 }
 
                 return Ok(false);